minor refactoring and renaming
