+++ src/miranda/convert/_reconstruction.py
@@ -15,10 +15,9 @@ from miranda.gis import subset_domain
 from miranda.io.utils import delayed_write, get_chunks_on_disk
 from miranda.scripting import LOGGING_CONFIG
 from miranda.utils import chunk_iterables
-from miranda.vocabularies import project_institute
+from miranda.vocabularies import project_institute, xarray_frequencies_to_cmip6like
 
 from ._aggregation import aggregate as aggregate_func
-from ._data_definitions import xarray_frequencies_to_cmip6like
 from .corrections import dataset_corrections
 
 logging.config.dictConfig(LOGGING_CONFIG)
