+++ src/miranda/vocabularies/__init__.py
@@ -4,3 +4,4 @@ from __future__ import annotations
 
 from . import eccc
 from ._projects import *
+from ._time import *
